Create Lichess-Omega
Remember:

This script will auto-move your chess pieces in Lichess with an accuracy range between 80-85%. The code works by using a set to keep track of the moves that have already been played, and then filtering out those moves from the list of valid moves to choose from. The threshold variable controls the accuracy range, and is set to 0.85 (85%) by default.